(fix) Update folder structure
