+++ tests/test_building.py
@@ -7,8 +7,9 @@ parentdir = os.path.dirname(currentdir)
 sys.path.insert(0, parentdir)
 
 import unittest
-from main.main import Building
-from model.model import Person, Staff, Fellow, Room, Office, LivingSpace
+from models.building import Building
+from models.person import Person, Staff, Fellow
+from models.room import Room, Office, LivingSpace
 
 living_space_names = ['Rm1', 'Rm2', 'Rm3', 'Rm4', 'Rm5', 'Rm6', 'Rm7', 'Rm8', 'Rm9', 'Rm10']
 office_names = ['Hogwarts', 'Valhalla', 'Oculus', 'Krypton', 'Shire', 'Narnia', 'Camelot', 'Mordor', 'Round Table', 'Midgar']
